Added the ability to add patrons
Added a new page for adding patrons, added the option in the index.php
file.  It displays an error when the user doesn't enter a name and a
confirmation when a patron is entered but I couldn't get it to fail the
insertion to see if the error properly displays there, but I assume it
does.